requirements.txt
datetime
numpy
playwright
tls-client2
redis
bezier
requests
g4f
pyjwt
flask
httpx